Move PeptideCountProcessor to worker thread
